Extract macOS browser rendering components
