+++ Sources/MacApp/Browser/BrowserTextPreview.swift
@@ -1,381 +1,8 @@
 import AppKit
-import ClipKittyAppleServices
-import ClipKittyMacPlatform
 import ClipKittyRust
-import ClipKittyShared
 import ObjectiveC.runtime
-import os.signpost
 import STTextKitPlus
 import SwiftUI
-import UniformTypeIdentifiers
-
-/// Max time to show stale content before clearing to spinner during slow loads.
-/// Used for both preview item loading and search result loading.
-private let staleContentTimeout: Duration = .milliseconds(150)
-private let poi = OSLog(subsystem: "com.eviljuliette.clipkitty", category: .pointsOfInterest)
-
-private enum SpinnerState: Equatable {
-    case idle
-    case debouncing(task: Task<Void, Never>)
-    case visible
-
-    static func == (lhs: SpinnerState, rhs: SpinnerState) -> Bool {
-        switch (lhs, rhs) {
-        case (.idle, .idle), (.visible, .visible), (.debouncing, .debouncing):
-            return true
-        default:
-            return false
-        }
-    }
-
-    mutating func cancel() {
-        if case let .debouncing(task) = self {
-            task.cancel()
-        }
-        self = .idle
-    }
-}
-
-private extension View {
-    @ViewBuilder
-    func clipKittyGlassBackground() -> some View {
-        if #available(macOS 26.0, *) {
-            glassEffect(.regular.interactive(), in: .rect)
-        } else {
-            background(.regularMaterial)
-        }
-    }
-}
-
-// MARK: - Subtle Hover Effect
-
-/// A view modifier that adds a subtle animated hover background effect.
-/// Use on button labels to provide visual feedback on hover.
-struct SubtleHoverEffect: ViewModifier {
-    let cornerRadius: CGFloat
-    let useCapsule: Bool
-    @State private var isHovered = false
-
-    init(cornerRadius: CGFloat = 9, useCapsule: Bool = false) {
-        self.cornerRadius = cornerRadius
-        self.useCapsule = useCapsule
-    }
-
-    func body(content: Content) -> some View {
-        content
-            .background {
-                if useCapsule {
-                    Capsule().fill(isHovered ? Color.primary.opacity(0.04) : Color.clear)
-                } else {
-                    RoundedRectangle(cornerRadius: cornerRadius)
-                        .fill(isHovered ? Color.primary.opacity(0.04) : Color.clear)
-                }
-            }
-            .contentShape(useCapsule ? AnyShape(Capsule()) : AnyShape(RoundedRectangle(cornerRadius: cornerRadius)))
-            .animation(.easeInOut(duration: 0.15), value: isHovered)
-            .onHover { isHovered = $0 }
-    }
-}
-
-extension View {
-    /// Adds a subtle hover background effect with rounded corners.
-    func subtleHover(cornerRadius: CGFloat = 9) -> some View {
-        modifier(SubtleHoverEffect(cornerRadius: cornerRadius))
-    }
-
-    /// Adds a subtle hover background effect with capsule shape.
-    func subtleHoverCapsule() -> some View {
-        modifier(SubtleHoverEffect(useCapsule: true))
-    }
-}
-
-// MARK: - File Preview
-
-@MainActor
-private enum FilePreviewIconCache {
-    private static let icons: NSCache<NSString, NSImage> = {
-        let cache = NSCache<NSString, NSImage>()
-        cache.countLimit = 256
-        return cache
-    }()
-
-    static func icon(forFile path: String) -> NSImage {
-        if let icon = icons.object(forKey: path as NSString) {
-            return icon
-        }
-        let icon = NSWorkspace.shared.icon(forFile: path)
-        icons.setObject(icon, forKey: path as NSString)
-        return icon
-    }
-}
-
-@MainActor
-private enum FileTextPreviewHighlightCache {
-    private static var highlightsByKey: [String: [Utf16HighlightRange]] = [:]
-
-    static func highlights(
-        forPreviewId previewId: String,
-        sample: String,
-        queryWords: [String]
-    ) -> [Utf16HighlightRange] {
-        guard !queryWords.isEmpty else { return [] }
-
-        let key = "\(previewId)|\(sample.utf16.count)|\(queryWords.joined(separator: "\u{1F}"))"
-        if let cached = highlightsByKey[key] {
-            return cached
-        }
-
-        let highlights = HighlightStyler.exactHighlights(in: sample, queryWords: queryWords)
-        highlightsByKey[key] = highlights
-        return highlights
-    }
-}
-
-private final class FilePreviewImageCache {
-    static let shared = FilePreviewImageCache()
-
-    private let cache: NSCache<NSString, NSImage> = {
-        let cache = NSCache<NSString, NSImage>()
-        cache.countLimit = 32
-        return cache
-    }()
-
-    func image(forKey key: String) -> NSImage? {
-        cache.object(forKey: key as NSString)
-    }
-
-    func setImage(_ image: NSImage, forKey key: String) {
-        cache.setObject(image, forKey: key as NSString)
-    }
-}
-
-private struct FilePreviewImageView<Fallback: View>: View {
-    private enum DecodeState {
-        case loading
-        case decoded(NSImage)
-        case failed
-    }
-
-    let cacheKey: String
-    let previewData: Data
-    let fallback: () -> Fallback
-
-    @State private var decodeState: DecodeState = .loading
-
-    var body: some View {
-        Group {
-            switch decodeState {
-            case .loading:
-                ProgressView()
-            case let .decoded(image):
-                Image(nsImage: image)
-                    .resizable()
-                    .aspectRatio(contentMode: .fit)
-                    .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-                    .padding(16)
-            case .failed:
-                fallback()
-            }
-        }
-        .task(id: cacheKey) {
-            if let cached = FilePreviewImageCache.shared.image(forKey: cacheKey) {
-                decodeState = .decoded(cached)
-                return
-            }
-
-            decodeState = .loading
-            let decoded = await Task.detached(priority: .userInitiated) { [previewData] in
-                NSImage(data: previewData)
-            }.value
-            guard !Task.isCancelled else { return }
-            guard let decoded else {
-                decodeState = .failed
-                return
-            }
-            FilePreviewImageCache.shared.setImage(decoded, forKey: cacheKey)
-            decodeState = .decoded(decoded)
-        }
-    }
-}
-
-struct FilePreviewView: View {
-    let itemId: String
-    let files: [FileEntry]
-    var searchQuery: String = ""
-
-    @ObservedObject private var settings = AppSettings.shared
-
-    /// Query words for highlighting (lowercased, non-empty)
-    private var queryWords: [String] {
-        searchQuery.lowercased()
-            .split(whereSeparator: { !$0.isLetter && !$0.isNumber })
-            .map(String.init)
-            .filter { !$0.isEmpty }
-    }
-
-    var body: some View {
-        if let previewFile {
-            switch previewFile.preview {
-            case let .text(text):
-                fileTextPreview(file: previewFile, text: text)
-            case let .image(previewData):
-                fileImagePreview(file: previewFile, previewData: previewData)
-            case .unavailable:
-                fileList
-            }
-        } else {
-            fileList
-        }
-    }
-
-    private var previewFile: FileEntry? {
-        for file in files {
-            switch file.preview {
-            case .text, .image:
-                return file
-            case .unavailable:
-                continue
-            }
-        }
-        return nil
-    }
-
-    private var fileList: some View {
-        ScrollView(.vertical, showsIndicators: true) {
-            VStack(spacing: 0) {
-                ForEach(Array(files.enumerated()), id: \.offset) { offset, file in
-                    fileRow(file)
-                    if offset != files.indices.last {
-                        Divider().padding(.leading, 52)
-                    }
-                }
-            }
-            .padding(.vertical, 12)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    private func previewHeader(file: FileEntry) -> some View {
-        HStack(spacing: 12) {
-            Image(nsImage: FilePreviewIconCache.icon(forFile: file.path))
-                .resizable()
-                .frame(width: 32, height: 32)
-
-            VStack(alignment: .leading, spacing: 3) {
-                highlightedFileText(file.filename, font: .system(size: 13, weight: .medium), color: .primary)
-                    .lineLimit(1)
-                highlightedFileText(file.path, font: .system(size: 11), color: .secondary)
-                    .lineLimit(1)
-                    .truncationMode(.middle)
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 10)
-        .background(.black.opacity(0.035))
-    }
-
-    private func fileTextPreview(file: FileEntry, text: FileTextPreviewSnapshot) -> some View {
-        let sample: String
-        switch text {
-        case let .complete(value):
-            sample = value
-        case let .truncated(value):
-            sample = value
-        }
-
-        let previewId = "\(itemId):\(file.path)"
-        let highlights = FileTextPreviewHighlightCache.highlights(
-            forPreviewId: previewId,
-            sample: sample,
-            queryWords: queryWords
-        )
-        let _ = { TextPreviewView.textCache[previewId] = sample }()
-
-        return VStack(spacing: 0) {
-            previewHeader(file: file)
-            Divider()
-            TextPreviewView(
-                itemId: previewId,
-                fontName: settings.previewFontName,
-                fontSize: settings.previewFontSize(12),
-                highlights: highlights,
-                scrollBehavior: .autoScroll,
-                interaction: .readOnly
-            )
-            switch text {
-            case .complete:
-                EmptyView()
-            case .truncated:
-                Divider()
-                Text("Preview truncated")
-                    .font(settings.appFont(size: 11))
-                    .foregroundStyle(.secondary)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.horizontal, 16)
-                    .padding(.vertical, 6)
-                    .background(.black.opacity(0.035))
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    private func fileImagePreview(file: FileEntry, previewData: Data) -> some View {
-        VStack(spacing: 0) {
-            previewHeader(file: file)
-            Divider()
-            ZStack(alignment: .topLeading) {
-                Color.black.opacity(0.04)
-                FilePreviewImageView(
-                    cacheKey: "\(itemId):\(file.path):\(previewData.count)",
-                    previewData: previewData
-                ) {
-                    fileList
-                }
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity, alignment: .topLeading)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    private func fileRow(_ file: FileEntry) -> some View {
-        HStack(spacing: 12) {
-            Image(nsImage: FilePreviewIconCache.icon(forFile: file.path))
-                .resizable()
-                .frame(width: 40, height: 40)
-
-            VStack(alignment: .leading, spacing: 4) {
-                highlightedFileText(file.filename, font: .system(size: 14, weight: .medium), color: .primary)
-                    .lineLimit(1)
-
-                highlightedFileText(file.path, font: .system(size: 11), color: .secondary)
-                    .lineLimit(1)
-                    .truncationMode(.middle)
-
-                if file.fileSize > 0 {
-                    Text(Utilities.formatBytes(Int64(file.fileSize)))
-                        .font(.system(size: 11))
-                        .foregroundStyle(.tertiary)
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-        }
-        .padding(.horizontal, 16)
-        .padding(.vertical, 8)
-    }
-
-    /// Highlight query word matches in file text
-    private func highlightedFileText(_ text: String, font: Font, color: Color) -> Text {
-        let highlights = HighlightStyler.exactHighlights(in: text, queryWords: queryWords)
-        guard !highlights.isEmpty else {
-            return Text(text).font(font).foregroundColor(color)
-        }
-
-        return Text(HighlightStyler.attributedText(text, highlights: highlights))
-            .font(font)
-            .foregroundColor(color)
-    }
-}
 
 // MARK: - Text Preview (AppKit)
 
@@ -1227,7 +854,7 @@ struct TextPreviewView: NSViewRepresentable {
 
         for matchRange in matchRanges {
             textLayoutManager.setRenderingAttributes(
-                HighlightStyler.renderingAttributes(for: matchRange.kind),
+                HighlightStyler.attributes(for: matchRange.kind),
                 for: matchRange.range
             )
         }
@@ -1406,704 +1033,6 @@ struct TextPreviewView: NSViewRepresentable {
             let text = textView.string
             lastRenderedText = text
             onTextChange(text)
-        }
-    }
-}
-
-// MARK: - Item Row
-
-@MainActor
-private enum RowIconCache {
-    private static let workspace = NSWorkspace.shared
-    private static let browserIcon: NSImage = {
-        if let browserURL = URL(string: "https://").flatMap({ workspace.urlForApplication(toOpen: $0) }) {
-            return workspace.icon(forFile: browserURL.path)
-        }
-        return workspace.icon(for: IconType.link.utType)
-    }()
-
-    private static let finderIcon: NSImage = {
-        if let finderURL = workspace.urlForApplication(withBundleIdentifier: "com.apple.finder") {
-            return workspace.icon(forFile: finderURL.path)
-        }
-        return workspace.icon(for: IconType.file.utType)
-    }()
-
-    private static var symbolIcons: [IconType: NSImage] = [:]
-    private static var sourceAppIcons: [String: NSImage] = [:]
-    private static var missingSourceAppBundleIDs: Set<String> = []
-
-    static func symbolImage(for iconType: IconType) -> NSImage {
-        if let cachedImage = symbolIcons[iconType] {
-            return cachedImage
-        }
-
-        let image: NSImage
-        switch iconType {
-        case .link:
-            image = browserIcon
-        case .file:
-            image = finderIcon
-        case .text, .image, .color:
-            image = workspace.icon(for: iconType.utType)
-        }
-
-        symbolIcons[iconType] = image
-        return image
-    }
-
-    static func sourceAppImage(bundleID: String) -> NSImage? {
-        if let cachedImage = sourceAppIcons[bundleID] {
-            return cachedImage
-        }
-        if missingSourceAppBundleIDs.contains(bundleID) {
-            return nil
-        }
-        guard let appURL = workspace.urlForApplication(withBundleIdentifier: bundleID) else {
-            missingSourceAppBundleIDs.insert(bundleID)
-            return nil
-        }
-
-        let image = workspace.icon(forFile: appURL.path)
-        sourceAppIcons[bundleID] = image
-        return image
-    }
-}
-
-private enum RowThumbnailCache {
-    private static let cache: NSCache<NSString, NSImage> = {
-        let cache = NSCache<NSString, NSImage>()
-        cache.countLimit = 128
-        cache.totalCostLimit = 16 * 1024 * 1024
-        return cache
-    }()
-
-    static func key(itemId: String, data: Data) -> String {
-        var hasher = Hasher()
-        hasher.combine(itemId)
-        hasher.combine(data.count)
-        data.withUnsafeBytes { rawBuffer in
-            let bytes = rawBuffer.bindMemory(to: UInt8.self)
-            for byte in bytes.prefix(16) {
-                hasher.combine(byte)
-            }
-            if bytes.count > 16 {
-                for byte in bytes.suffix(16) {
-                    hasher.combine(byte)
-                }
-            }
-        }
-        return "\(itemId)-\(data.count)-\(hasher.finalize())"
-    }
-
-    static func image(forKey key: String) -> NSImage? {
-        cache.object(forKey: key as NSString)
-    }
-
-    static func setImage(_ image: NSImage, forKey key: String, cost: Int) {
-        cache.setObject(image, forKey: key as NSString, cost: cost)
-    }
-}
-
-private struct RowThumbnailView: View {
-    let itemId: String
-    let data: Data
-
-    @State private var decodedImage: NSImage?
-
-    private var cacheKey: String {
-        RowThumbnailCache.key(itemId: itemId, data: data)
-    }
-
-    var body: some View {
-        Group {
-            if let image = decodedImage ?? RowThumbnailCache.image(forKey: cacheKey) {
-                Image(nsImage: image)
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-            } else {
-                Image(systemName: "photo")
-                    .resizable()
-            }
-        }
-        .task(id: cacheKey) {
-            await decodeImage(cacheKey: cacheKey, data: data)
-        }
-    }
-
-    @MainActor
-    private func decodeImage(cacheKey: String, data: Data) async {
-        if let cachedImage = RowThumbnailCache.image(forKey: cacheKey) {
-            decodedImage = cachedImage
-            return
-        }
-
-        decodedImage = nil
-        let image = await Task.detached(priority: .utility) { [data] in
-            NSImage(data: data)
-        }.value
-        guard !Task.isCancelled, let image else { return }
-        RowThumbnailCache.setImage(image, forKey: cacheKey, cost: data.count)
-        decodedImage = image
-    }
-}
-
-struct ItemRow: View {
-    let metadata: ItemMetadata
-    let presentation: RowPresentation
-    let isSelected: Bool
-    let isContextMenuTargeted: Bool
-    let hasUserNavigated: Bool
-    let hasPendingEdit: Bool
-    let onTap: () -> Void
-    let contextMenuActions: [BrowserActionItem]
-    let onContextMenuAction: (BrowserActionItem) -> Void
-    let onContextMenuDelete: () -> Void
-    let onContextMenuShow: () -> Void
-    let onContextMenuHide: () -> Void
-
-    @ObservedObject private var settings = AppSettings.shared
-
-    private var accentSelected: Bool {
-        isSelected && hasUserNavigated && !hasPendingEdit
-    }
-
-    /// Height for exactly 1 line of text, scaled with text size setting
-    private var rowHeight: CGFloat {
-        settings.scaled(32)
-    }
-
-    // MARK: - Display Text (Simplified - SwiftUI handles truncation)
-
-    private var displayExcerpt: (text: String, highlights: [Utf16HighlightRange], lineNumber: UInt64?) {
-        switch presentation {
-        case let .baseline(excerpt):
-            return (excerpt.text, [], nil)
-        case let .matched(excerpt):
-            return (excerpt.text, excerpt.highlights, excerpt.lineNumber)
-        case let .deferred(_, placeholder):
-            switch placeholder {
-            case let .baseline(excerpt), let .provisional(excerpt):
-                return (excerpt.text, [], nil)
-            case let .compatibleCached(_, excerpt):
-                return (excerpt.text, excerpt.highlights, excerpt.lineNumber)
-            }
-        case let .unavailable(fallback, _):
-            return (fallback.text, [], nil)
-        }
-    }
-
-    private var showsSourceAppBadge: Bool {
-        switch metadata.icon {
-        case let .symbol(iconType):
-            return iconType != .link && iconType != .file
-        case .thumbnail, .colorSwatch:
-            return true
-        }
-    }
-
-    private var lineNumberFont: Font {
-        switch settings.fontPreference {
-        case .iosevkaCharon:
-            return settings.previewFont(size: 13)
-        case .system:
-            return settings.appFont(size: 13)
-        }
-    }
-
-    var body: some View {
-        // 1. Wrap the content inside a Button
-        Button(action: onTap) {
-            HStack(spacing: 6) {
-                // Content type icon with badge overlay (or pencil when editing)
-                Group {
-                    if hasPendingEdit {
-                        // Show pencil emoji when item has pending edit
-                        Text("✏️")
-                            .font(.system(size: settings.scaled(24)))
-                            .frame(width: settings.scaled(32), height: settings.scaled(32))
-                    } else {
-                        ZStack(alignment: .bottomTrailing) {
-                            // Main icon: image thumbnail, browser icon for links, color swatch, or SF symbol
-                            Group {
-                                switch metadata.icon {
-                                case let .thumbnail(bytes):
-                                    RowThumbnailView(itemId: metadata.itemId, data: bytes)
-                                case let .colorSwatch(rgba):
-                                    RoundedRectangle(cornerRadius: 6)
-                                        .fill(Color(nsColor: NSColor(
-                                            red: CGFloat((rgba >> 24) & 0xFF) / 255.0,
-                                            green: CGFloat((rgba >> 16) & 0xFF) / 255.0,
-                                            blue: CGFloat((rgba >> 8) & 0xFF) / 255.0,
-                                            alpha: CGFloat(rgba & 0xFF) / 255.0
-                                        )))
-                                        .overlay(
-                                            RoundedRectangle(cornerRadius: 6)
-                                                .strokeBorder(Color.primary.opacity(0.15), lineWidth: 1)
-                                        )
-                                case let .symbol(iconType):
-                                    Image(nsImage: RowIconCache.symbolImage(for: iconType))
-                                        .resizable()
-                                }
-                            }
-                            .frame(width: settings.scaled(32), height: settings.scaled(32))
-                            .clipShape(RoundedRectangle(cornerRadius: 4))
-
-                            // Badge: Bookmark icon for bookmarked items, otherwise source app icon
-                            if metadata.tags.contains(.bookmark) {
-                                Image("BookmarkIcon")
-                                    .resizable()
-                                    .frame(width: 18, height: 18)
-                                    .shadow(color: .black.opacity(0.3), radius: 1, x: 0, y: 1)
-                                    .offset(x: 4, y: 4)
-                                    .transition(.scale.combined(with: .opacity))
-                            } else if let bundleID = metadata.sourceAppBundleId,
-                                      let sourceAppImage = RowIconCache.sourceAppImage(bundleID: bundleID)
-                            {
-                                if showsSourceAppBadge {
-                                    Image(nsImage: sourceAppImage)
-                                        .resizable()
-                                        .frame(width: 22, height: 22)
-                                        .clipShape(RoundedRectangle(cornerRadius: 3))
-                                        .shadow(color: .black.opacity(0.3), radius: 1, x: 0, y: 1)
-                                        .offset(x: 4, y: 4)
-                                }
-                            }
-                        }
-                    }
-                }
-                .frame(width: settings.scaled(38), height: settings.scaled(38))
-                .allowsHitTesting(false)
-
-                // Line number (shown in search mode when line > 1)
-                if let lineNumber = displayExcerpt.lineNumber, lineNumber > 1 {
-                    Text("L\(lineNumber):")
-                        .font(lineNumberFont)
-                        .foregroundColor(accentSelected ? .white.opacity(0.7) : .secondary)
-                        .lineLimit(1)
-                        .fixedSize()
-                        .allowsHitTesting(false)
-                }
-
-                // Text content - SwiftUI Three-Part HStack with layout priorities
-                HStack(spacing: 6) {
-                    HighlightedTextView(
-                        text: displayExcerpt.text,
-                        highlights: displayExcerpt.highlights,
-                        accentSelected: accentSelected,
-                        textScale: settings.textScale,
-                        fontPreference: settings.fontPreference
-                    )
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .allowsHitTesting(false)
-                    .layoutPriority(1)
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .frame(maxWidth: .infinity, minHeight: rowHeight, maxHeight: rowHeight, alignment: .leading)
-            .padding(.horizontal, 4)
-            .padding(.vertical, 4)
-            .background {
-                if isSelected && hasUserNavigated && hasPendingEdit {
-                    // Editing state: darker grey background
-                    Color.primary.opacity(0.35)
-                } else if accentSelected {
-                    Color.selectionBackground
-                } else if isContextMenuTargeted && !isSelected {
-                    Color.primary.opacity(0.11)
-                } else if isSelected {
-                    Color.primary.opacity(0.225)
-                } else {
-                    Color.clear
-                }
-            }
-            .overlay {
-                if isContextMenuTargeted {
-                    RoundedRectangle(cornerRadius: 10, style: .continuous)
-                        .strokeBorder(Color.primary.opacity(0.22), lineWidth: 1)
-                }
-            }
-            // Each keyed to its own state so selection changes stay instant:
-            // the pencil/background crossfade on entering edit mode, the
-            // context-target ring/fill fade, and the bookmark badge pop.
-            .animation(.easeInOut(duration: 0.15), value: hasPendingEdit)
-            .animation(.easeOut(duration: 0.12), value: isContextMenuTargeted)
-            .animation(.spring(response: 0.3, dampingFraction: 0.55), value: metadata.tags.contains(.bookmark))
-            .clipShape(RoundedRectangle(cornerRadius: 10, style: .continuous))
-            .contentShape(Rectangle())
-        }
-        // 2. Apply the plain style so it behaves like a standard row instead of a system button
-        .buttonStyle(.plain)
-        .overlay {
-            RightClickPopoverOverlay(
-                actions: contextMenuActions,
-                onShow: onContextMenuShow,
-                onHide: onContextMenuHide,
-                onAction: onContextMenuAction,
-                onConfirmDelete: onContextMenuDelete
-            )
-        }
-        .accessibilityElement(children: .combine)
-        .accessibilityLabel(displayExcerpt.text)
-        .accessibilityHint(AppSettings.shared.pasteMode == .autoPaste ? String(localized: "Double tap to paste") : String(localized: "Double tap to copy"))
-        .accessibilityAddTraits(.isButton)
-        .accessibilityAddTraits(isSelected ? .isSelected : [])
-    }
-}
-
-// MARK: - Right-Click Popover
-
-struct RightClickPopoverOverlay: NSViewRepresentable {
-    let actions: [BrowserActionItem]
-    let onShow: () -> Void
-    let onHide: () -> Void
-    let onAction: (BrowserActionItem) -> Void
-    let onConfirmDelete: () -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator()
-    }
-
-    func makeNSView(context: Context) -> RightClickView {
-        let view = RightClickView()
-        view.coordinator = context.coordinator
-        context.coordinator.actions = actions
-        context.coordinator.onShow = onShow
-        context.coordinator.onHide = onHide
-        context.coordinator.onAction = onAction
-        context.coordinator.onConfirmDelete = onConfirmDelete
-        return view
-    }
-
-    func updateNSView(_: RightClickView, context: Context) {
-        context.coordinator.actions = actions
-        context.coordinator.onShow = onShow
-        context.coordinator.onHide = onHide
-        context.coordinator.onAction = onAction
-        context.coordinator.onConfirmDelete = onConfirmDelete
-    }
-
-    @MainActor
-    final class Coordinator {
-        var actions: [BrowserActionItem] = []
-        var onShow: (() -> Void)?
-        var onHide: (() -> Void)?
-        var onAction: ((BrowserActionItem) -> Void)?
-        var onConfirmDelete: (() -> Void)?
-        private var activeMenuHandler: MenuActionHandler?
-
-        deinit {
-            activeMenuHandler = nil
-        }
-
-        func makeMenu() -> NSMenu {
-            let menu = NSMenu()
-            menu.autoenablesItems = false
-
-            let handler = MenuActionHandler(
-                onAction: { [weak self] action in
-                    self?.onAction?(action)
-                },
-                onConfirmDelete: { [weak self] in
-                    self?.onConfirmDelete?()
-                }
-            )
-            activeMenuHandler = handler
-
-            for (index, action) in actions.enumerated() {
-                if BrowserActionItem.showsDivider(before: index, in: actions) {
-                    menu.addItem(.separator())
-                }
-
-                let item = NSMenuItem(
-                    title: action.label,
-                    action: #selector(MenuActionHandler.handleMenuItem(_:)),
-                    keyEquivalent: ""
-                )
-                item.target = handler
-                item.representedObject = action
-                item.image = NSImage(
-                    systemSymbolName: action.systemImageName,
-                    accessibilityDescription: action.label
-                )
-                menu.addItem(item)
-            }
-
-            return menu
-        }
-    }
-
-    final class RightClickView: NSView {
-        weak var coordinator: Coordinator?
-
-        override func hitTest(_ point: NSPoint) -> NSView? {
-            guard shouldHandleCurrentEvent else { return nil }
-            return super.hitTest(point)
-        }
-
-        override func mouseDown(with event: NSEvent) {
-            if event.modifierFlags.contains(.control) {
-                rightMouseDown(with: event)
-                return
-            }
-            super.mouseDown(with: event)
-        }
-
-        override func rightMouseDown(with event: NSEvent) {
-            guard let coordinator, !coordinator.actions.isEmpty else { return }
-
-            coordinator.onShow?()
-            let menu = coordinator.makeMenu()
-            let clickPoint = convert(event.locationInWindow, from: nil)
-            menu.popUp(positioning: nil, at: clickPoint, in: self)
-            coordinator.onHide?()
-        }
-
-        override func menu(for _: NSEvent) -> NSMenu? {
-            nil
-        }
-
-        private var shouldHandleCurrentEvent: Bool {
-            guard let event = NSApp.currentEvent else { return false }
-            switch event.type {
-            case .rightMouseDown, .rightMouseUp:
-                return true
-            case .leftMouseDown, .leftMouseUp:
-                return event.modifierFlags.contains(.control)
-            default:
-                return false
-            }
-        }
-    }
-
-    @MainActor
-    private final class MenuActionHandler: NSObject {
-        let onAction: (BrowserActionItem) -> Void
-        let onConfirmDelete: () -> Void
-
-        init(onAction: @escaping (BrowserActionItem) -> Void, onConfirmDelete: @escaping () -> Void) {
-            self.onAction = onAction
-            self.onConfirmDelete = onConfirmDelete
-        }
-
-        @objc
-        func handleMenuItem(_ sender: NSMenuItem) {
-            guard let action = sender.representedObject as? BrowserActionItem else { return }
-            if action.isDestructive {
-                onConfirmDelete()
-            } else {
-                onAction(action)
-            }
-        }
-    }
-}
-
-// MARK: - Three-Part HStack Highlighted Text
-
-/// SwiftUI-native text view using Three-Part HStack strategy for search highlighting.
-/// Uses layout priorities to guarantee the first highlight is always visible while maximizing context.
-/// - Prefix: Truncates from head (`.head`) showing "...text"
-/// - Highlight: Has `.layoutPriority(1)` to claim space first, never pushed off-screen
-/// - Suffix: Truncates from tail (`.tail`) showing "text..."
-///
-/// Uses HighlightStyler for all index calculations with proper Unicode scalar handling.
-struct HighlightedTextView: View, Equatable {
-    let text: String
-    let highlights: [Utf16HighlightRange]
-    let accentSelected: Bool
-    let textScale: CGFloat
-    let fontPreference: AppFontPreference
-
-    private var textColor: Color {
-        accentSelected ? .white : .primary
-    }
-
-    private var font: Font {
-        let size = AppFontMetrics.size(15 * textScale, for: fontPreference)
-        switch fontPreference {
-        case .iosevkaCharon:
-            return .custom(FontManager.sansSerifName(for: .iosevkaCharon), size: size)
-        case .system:
-            return .system(size: size)
-        }
-    }
-
-    var body: some View {
-        // Use firstTextBaseline so text aligns perfectly even with different weights
-        HStack(alignment: .firstTextBaseline, spacing: 0) {
-            if let firstHighlight = highlights.first {
-                // Use HighlightStyler so the row renderer stays on UTF-16 offsets end to end.
-                let (prefix, match, suffix) = HighlightStyler.splitText(text, highlight: firstHighlight)
-                let suffixStartUtf16Offset = Int(firstHighlight.utf16End)
-
-                // 1. PREFIX: Truncates on the left ("...text")
-                if !prefix.isEmpty {
-                    Text(prefix)
-                        .font(font)
-                        .foregroundColor(textColor)
-                        .lineLimit(1)
-                        .truncationMode(.head)
-                }
-
-                // 2. HIGHLIGHT: High priority ensures it claims space first
-                highlightedMatchView(match: match, kind: firstHighlight.kind)
-
-                // 3. SUFFIX: Truncates on the right ("text...")
-                // Apply any additional highlights that fall within suffix
-                if !suffix.isEmpty {
-                    suffixView(suffix: suffix, suffixStartUtf16Offset: suffixStartUtf16Offset)
-                        .font(font)
-                        .foregroundColor(textColor)
-                        .lineLimit(1)
-                        .truncationMode(.tail)
-                }
-
-            } else {
-                // No highlights - simple text with tail truncation
-                Text(text)
-                    .font(font)
-                    .foregroundColor(textColor)
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-            }
-        }
-        // Ensure text aligns to the left if shorter than container
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-
-    /// Build the highlighted match view with optional underline
-    private func highlightedMatchView(match: String, kind: HighlightKind) -> some View {
-        Text(HighlightStyler.attributedFragment(match, kind: kind))
-            .font(font)
-            .foregroundColor(textColor)
-            .lineLimit(1)
-            .truncationMode(.tail)
-            .layoutPriority(1)
-    }
-
-    /// Build suffix view with any additional highlights
-    @ViewBuilder
-    private func suffixView(suffix: String, suffixStartUtf16Offset: Int) -> some View {
-        // Check for additional highlights in the suffix (beyond the first one)
-        let suffixUtf16Count = suffix.utf16.count
-        let additionalHighlights = highlights.dropFirst().filter { h in
-            HighlightStyler.highlightInSuffix(
-                h,
-                suffixStartUtf16Offset: suffixStartUtf16Offset,
-                suffixUtf16Count: suffixUtf16Count
-            )
-        }
-
-        if additionalHighlights.isEmpty {
-            Text(suffix)
-        } else {
-            Text(HighlightStyler.attributedSuffix(
-                suffix,
-                suffixStartUtf16Offset: suffixStartUtf16Offset,
-                highlights: Array(additionalHighlights)
-            ))
-        }
-    }
-}
-
-// MARK: - Action Option Row
-
-struct ActionOptionRow: View {
-    let label: String
-    let actionID: String
-    let systemImageName: String
-    var isHighlighted: Bool = false
-    var isDestructive: Bool = false
-    var onHover: ((Bool) -> Void)?
-    let action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            HStack(spacing: 8) {
-                Image(systemName: systemImageName)
-                    .font(.system(size: 13, weight: .medium))
-                    .frame(width: 14, alignment: .center)
-
-                Text(label)
-                    .font(.system(size: 13))
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-            .foregroundStyle(foregroundColor)
-            .padding(.horizontal, 8)
-            .padding(.vertical, 4)
-            .background {
-                RoundedRectangle(cornerRadius: 7)
-                    .fill(backgroundColor)
-            }
-            .animation(.easeOut(duration: 0.12), value: isHighlighted)
-            .contentShape(RoundedRectangle(cornerRadius: 7))
-        }
-        .buttonStyle(.plain)
-        .onHover { onHover?($0) }
-        .accessibilityIdentifier("Action_\(actionID)")
-    }
-
-    private var foregroundColor: Color {
-        if isHighlighted { return .white }
-        if isDestructive { return .red }
-        return .secondary
-    }
-
-    private var backgroundColor: Color {
-        if isHighlighted {
-            return isDestructive ? Color.red.opacity(0.8) : .selectionBackground
-        }
-        return Color.clear
-    }
-}
-
-// MARK: - Selection Background
-
-extension Color {
-    /// Selection background matching system tint, slightly desaturated for a subtler look.
-    static var selectionBackground: Color {
-        Color(nsColor: .selectedContentBackgroundColor.desaturated(by: 0.10))
-    }
-}
-
-private extension NSColor {
-    /// Returns a new color with saturation reduced by the given fraction (0.0–1.0).
-    func desaturated(by amount: CGFloat) -> NSColor {
-        guard let hsb = usingColorSpace(.sRGB) else { return self }
-        var h: CGFloat = 0, s: CGFloat = 0, b: CGFloat = 0, a: CGFloat = 0
-        hsb.getHue(&h, saturation: &s, brightness: &b, alpha: &a)
-        let newSaturation = max(0, s - amount)
-        return NSColor(hue: h, saturation: newSaturation, brightness: b, alpha: a)
-    }
-}
-
-// MARK: - Highlight Kind Color Mapping
-
-// MARK: - Hide Scroll Indicators When System Uses Overlay Style
-
-/// Hides scroll indicators when the system preference is "Show scroll bars: When scrolling" (overlay style).
-/// Detects scrolling via ScrollView geometry and shows indicators only while actively scrolling.
-/// This prevents the brief scrollbar flash when the panel appears.
-struct HideScrollIndicatorsWhenOverlay: ViewModifier {
-    let displayVersion: Int
-    @State private var hasScrolled = false
-
-    func body(content: Content) -> some View {
-        if #available(macOS 15.0, *), NSScroller.preferredScrollerStyle == .overlay {
-            content
-                .scrollIndicators(hasScrolled ? .automatic : .never)
-                .onScrollGeometryChange(for: CGFloat.self) { geometry in
-                    geometry.contentOffset.y
-                } action: { _, _ in
-                    if !hasScrolled {
-                        hasScrolled = true
-                    }
-                }
-                .onChange(of: displayVersion) { _, _ in
-                    hasScrolled = false
-                }
-        } else {
-            content
         }
     }
 }
